+++ tests/services/test_normalize_address.py
@@ -1,0 +1,5 @@
+"""Test the normalize address method of the ShipEngine SDK."""
+
+
+# class TestNormalizeAddress:
+#     pass
